+++ nevergrad/parametrization/choice.py
